+++ lib/router/router.dart
@@ -4,6 +4,7 @@ import 'package:youthetree/ui/test/test_code_pop_up.dart';
 import 'package:youthetree/ui/page/home_page.dart';
 import 'package:youthetree/ui/page/initial_test.dart';
 import 'package:youthetree/ui/page/splash_page.dart';
+import 'package:youthetree/ui/test/test_login_page.dart';
 import 'package:youthetree/ui/test/test_tree_updates.dart';
 import 'package:youthetree/ui/test/user_avatar_test.dart';
 import 'package:youthetree/ui/test/dynamic_column.dart';
@@ -30,6 +31,7 @@ class RouteName {
   static const String testUserAvatar = 'testUserAvatar';
   static const String testDynamic = 'testDynamicColumn';
   static const String testTreeUpdates = 'testTreeUpdates';
+  static const String testLoginPage = 'testLoginPage';
 
 
 }
@@ -40,4 +42,5 @@ Map<String, WidgetBuilder> testRoutes = {
   RouteName.testUserAvatar: (context) => UserAvatarTest(),
   RouteName.testDynamic: (context) => DynamicColumn(),
   RouteName.testTreeUpdates: (context) => TestTreeUpdates(),
+  RouteName.testLoginPage: (context) => TestLoginPage(),
 };
