+++ lib/ui/test/test_login_page.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+import 'package:youthetree/ui/page/login_page.dart';
+
+class TestLoginPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: LoginPage(),
+    );
+  }
+}
